refactored to accomodate future activities and trigger
