+++ pulsar/activity/publish/descriptor.json
