+++ pulsar/trigger/subscriber/descriptor.json
