+++ pulsar/trigger/subscriber/trigger_test.go
@@ -1,4 +1,4 @@
-package trigger
+package subscriber
 
 /*
 import (
